feat/gas-refund: decouple tx fees/gas refund indexing from merkle tree computing
